Start redesigning Creevey around Pipeline objects
