+++ creevey/util/constants.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+from typing import Union
+
+PathOrStr = Union[Path, str]
